+++ source/helpers/getPhoneInputWithCountryStateUpdateFromNewProps.js
@@ -127,7 +127,7 @@ export default function getPhoneInputWithCountryStateUpdateFromNewProps(props, p
 	// which happens in `this.onChange()` right after `this.setState()`.
 	// If this `getDerivedStateFromProps()` call isn't ignored
 	// then the country flag would reset on each input.
-	if (newValue !== prevValue && newValue !== value) {
+	if (!valuesAreEqual(newValue, prevValue) && !valuesAreEqual(newValue, value)) {
 		let phoneNumber
 		let parsedCountry
 		if (newValue) {
@@ -171,4 +171,27 @@ export default function getPhoneInputWithCountryStateUpdateFromNewProps(props, p
 	// `phoneDigits` didn't change, because `value` didn't change.
 	//
 	// So no need to update state.
+}
+
+function valuesAreEqual(value1, value2) {
+	// If `value` has been set to `null` externally then convert it to `undefined`.
+	//
+	// For example, `react-hook-form` sets `value` to `null` when the user clears the input.
+	// https://gitlab.com/catamphetamine/react-phone-number-input/-/issues/164
+	// In that case, without this conversion of `null` to `undefined`, it would reset
+	// the selected country to `defaultCountry` because in that case `newValue !== value`
+	// because `null !== undefined`.
+	//
+	// Historically, empty `value` is encoded as `undefined`.
+	// Perhaps empty `value` would be better encoded as `null` instead.
+	// But because that would be a potentially breaking change for some people,
+	// it's left as is for the current "major" version of this library.
+	//
+	if (value1 === null) {
+		value1 = undefined
+	}
+	if (value2 === null) {
+		value2 = undefined
+	}
+	return value1 === value2
 }